requirements.txt
pyTelegramBotAPI
fastapi
uvicorn
psycopg2-binary
python-dotenv
pandas